+++ fastfort/orm/sqlalchemy/__init__.py
@@ -6,7 +6,7 @@ from .adapter import SQLAlchemyAdapter
 from .backend import SQLAlchemyBackend, SQLAlchemyUnitOfWork
 from .dialects import DialectProfile, profile_for
 from .introspect import introspect_model, is_sqlalchemy_model
-from .models import ApiTokenMixin, SignInRecordMixin
+from .models import ApiTokenMixin, FavoriteMixin, SignInRecordMixin
 from .query import QueryBuilder
 from .types import Classification, TypeRule, classify, register_type
 
@@ -14,6 +14,7 @@ __all__ = [
     "ApiTokenMixin",
     "Classification",
     "DialectProfile",
+    "FavoriteMixin",
     "QueryBuilder",
     "SQLAlchemyAdapter",
     "SQLAlchemyBackend",